DefaultLeader - Bully Algorithm
Server which Serves user response. Applied Bully algorithm to select leader among other servers when any of them is down. Used Multicast to send the same request to all servers and get response from all of them and send the correct result to front end based on majority of success or Failure message.